requirements.txt
arcade==3.0.0.dev19
attrs==22.2.0
cffi==1.15.1
Pillow==9.4.0
pycapnp==1.3.0
pycparser==2.21
pyglet==2.0.5
pymunk==6.4.0
pytiled-parser==2.2.0
typing_extensions==4.5.0
mysql-connector-python~=8.0.32
